Added formatter for Slack.
Builds a slack message attachment JSON structure (https://api.slack.com/docs/attachments)
Minor changes had to be made to Formatter interface to accomodate this change. As a
result, Formatters now work on a slice of JSON-results from the different hosts, instead
of working with the result of one host at a time.